Medium-sized update
Okay, buttons are looking good. Hopefully the bolding + colors are a readable combination. The continued failure of the body element to have a BG-color is troubling. And so too is the question of what to do for the mobile ver of the app. That it isn't even appearing on the nav-bar, which is acting oddly itself, only makes this worse. 

Maybe it's time to completely delete the files and then restart again afresh. There seems to be some lingering traces of the Angular in the Visual Studio's autocomplete -- maybe that's the issue?